passed every damage and health point to int

+++ ProjectAnnihilation/Assets/Scripts/Interfaces/IDamageable.cs
@@ -5,13 +5,13 @@ using UnityEngine;
 public struct DamageData
 {
     public Vector3? knockback; // Can be null
-    public float damage;
+    public int damage;
     public float hitStun;
     public bool ignoreDefense;
 
     public DamageData(float damage,  float hitStun = 0f, Vector3? knockback = null, bool ignoreDefense = false)
     {
-        this.damage = damage;
+        this.damage = (int)damage;
         this.knockback = knockback;
         this.hitStun = hitStun;
         this.ignoreDefense = ignoreDefense;
@@ -22,5 +22,5 @@ public interface IDamageable
 {
     public void Damage(DamageData damageData, IDamageable from = null);
 
-    public void Heal(float heal);
+    public void Heal(int heal);
 }
